third_party/WebKit/LayoutTests/external/wpt/html/infrastructure/urls/dynamic-changes-to-base-urls/dynamic-urls.sub-expected.txt

This is a testharness.js-based test.
FAIL The 'href' attribute of the 'a' elementassert_equals: The 'href' attribute is incorrect. expected "http://www.web-platform.test:8001/test.txt" but got "http://web-platform.test:8001/html/infrastructure/urls/dynamic-changes-to-base-urls/test.txt"
FAIL The 'href' attribute of the 'link' elementassert_equals: The 'href' attribute is incorrect. expected "http://www.web-platform.test:8001/test.txt" but got "http://web-platform.test:8001/html/infrastructure/urls/dynamic-changes-to-base-urls/test.txt"
FAIL The 'href' attribute of the 'area' elementassert_equals: The 'href' attribute is incorrect. expected "http://www.web-platform.test:8001/test.txt" but got "http://web-platform.test:8001/html/infrastructure/urls/dynamic-changes-to-base-urls/test.txt"
FAIL The 'cite' attribute of the 'q' elementassert_equals: The 'cite' attribute is incorrect. expected "http://www.web-platform.test:8001/test.txt" but got "http://web-platform.test:8001/html/infrastructure/urls/dynamic-changes-to-base-urls/test.txt"
FAIL The 'cite' attribute of the 'blockquote' elementassert_equals: The 'cite' attribute is incorrect. expected "http://www.web-platform.test:8001/test.txt" but got "http://web-platform.test:8001/html/infrastructure/urls/dynamic-changes-to-base-urls/test.txt"
FAIL The 'cite' attribute of the 'ins' elementassert_equals: The 'cite' attribute is incorrect. expected "http://www.web-platform.test:8001/test.txt" but got "http://web-platform.test:8001/html/infrastructure/urls/dynamic-changes-to-base-urls/test.txt"
FAIL The 'cite' attribute of the 'del' elementassert_equals: The 'cite' attribute is incorrect. expected "http://www.web-platform.test:8001/test.txt" but got "http://web-platform.test:8001/html/infrastructure/urls/dynamic-changes-to-base-urls/test.txt"
FAIL The 'src' attribute of the 'img' elementassert_equals: The 'src' attribute is incorrect. expected "http://www.web-platform.test:8001/test.txt" but got "http://web-platform.test:8001/html/infrastructure/urls/dynamic-changes-to-base-urls/test.txt"
FAIL The 'src' attribute of the 'embed' elementassert_equals: The 'src' attribute is incorrect. expected "http://www.web-platform.test:8001/test.txt" but got "http://web-platform.test:8001/html/infrastructure/urls/dynamic-changes-to-base-urls/test.txt"
FAIL The 'src' attribute of the 'video' elementassert_equals: The 'src' attribute is incorrect. expected "http://www.web-platform.test:8001/test.txt" but got "http://web-platform.test:8001/html/infrastructure/urls/dynamic-changes-to-base-urls/test.txt"
FAIL The 'src' attribute of the 'iframe' elementassert_equals: The 'src' attribute is incorrect. expected "http://www.web-platform.test:8001/test.txt" but got "http://web-platform.test:8001/html/infrastructure/urls/dynamic-changes-to-base-urls/test.txt"
FAIL The 'src' attribute of the 'script' elementassert_equals: The 'src' attribute is incorrect. expected "http://www.web-platform.test:8001/test.txt" but got "http://web-platform.test:8001/html/infrastructure/urls/dynamic-changes-to-base-urls/test.txt"
FAIL The 'src' attribute of the 'source' elementassert_equals: The 'src' attribute is incorrect. expected "http://www.web-platform.test:8001/test.txt" but got "http://web-platform.test:8001/html/infrastructure/urls/dynamic-changes-to-base-urls/test.txt"
FAIL The 'src' attribute of the 'track' elementassert_equals: The 'src' attribute is incorrect. expected "http://www.web-platform.test:8001/test.txt" but got "http://web-platform.test:8001/html/infrastructure/urls/dynamic-changes-to-base-urls/test.txt"
FAIL The 'action' attribute of the 'form' elementassert_equals: The 'action' attribute is incorrect. expected "http://www.web-platform.test:8001/test.txt" but got "http://web-platform.test:8001/html/infrastructure/urls/dynamic-changes-to-base-urls/test.txt"
FAIL The 'data' attribute of the 'object' elementassert_equals: The 'data' attribute is incorrect. expected "http://www.web-platform.test:8001/test.txt" but got "http://web-platform.test:8001/html/infrastructure/urls/dynamic-changes-to-base-urls/test.txt"
FAIL The 'formaction' attribute of the 'button' elementassert_equals: The 'formAction' attribute is incorrect. expected "http://www.web-platform.test:8001/test.txt" but got "http://web-platform.test:8001/html/infrastructure/urls/dynamic-changes-to-base-urls/test.txt"
FAIL Change the base URL must effect the descendant elements onlyassert_not_equals: The inner element must be effected. got disallowed value "http://web-platform.test:8001/html/infrastructure/urls/dynamic-changes-to-base-urls/test.txt"
Harness: the test ran to completion.
